Disable mininification + import changes
Import changes should reduce the size of the bundle a little.
Particularly, not getting the entirety of lodash.
